automation/requirements.txt
requests
feedparser
python-slugify
python-dotenv
Pillow
groq